Added loop for each connection
You need to keep the connection alive "forever" (for now, or until you get some signal that the connection should be killed). This allows us to use the same websocket connection over and over again instead of initializing one for every new message to the server.

An alternative option with more overhead but less theoretical risks is to handle each message as a new websocket connection. But we will have to change the Minecraft_API for this to work.