Added multiple data objects for Minecraft_Store and Fixed File Writes
I capped the buffer on the append file to request that it flushes after 1000 lines. This is not a huge change, but is better than having n*8000 possible lines in memory.

I also fixed the Data and Minecraft_Store objects so they append properly so you have multiple files that can be written to at once. This fixes the renaming issue